Started step 6
I just ran out of time and energy. That is on me. Which is why I fought for the one I actually did do. My kids have been stuck at home for too long, so I couldn't get the time to just work on my homework. Which is my problem. 10/70